Fix getting stuck in loop in socket_listen
If `socket_listen` was called, and a connection was made right
away, it was possible that the state of the socket would change
to `SNSR_SOCK_ESTABLISHED` before the check ever saw it enter
`SNSR_SOCK_LISTEN`, and the code would be stuck in an endless
loop.

This fixes the issue by accepting both `SNSR_SOCK_LISTEN` and
`SNSR_SOCK_ESTABLISHED` as valid states to continue.